extension point work, separated maven UI via extension point methods and APIs.
