+++ google-cloud-apis/resources/META-INF/google-cloud-apis.xml
@@ -21,6 +21,11 @@
         <extensionPoint qualifiedName="com.google.gct.cloudapis.cloudApiRunConfiguration"
                         interface="com.google.cloud.tools.intellij.cloudapis.CloudApiRunConfigurationProvider"/>
     </extensionPoints>
+    <extensionPoints>
+        <extensionPoint qualifiedName="com.google.gct.cloudapis.cloudApiUiExtension"
+                        interface="com.google.cloud.tools.intellij.cloudapis.CloudApiUiExtension"/>
+    </extensionPoints>
+
     <actions>
         <group id="GoogleCloudTools.cloudapis">
             <action id="GoogleCloudTools.AddCloudLibraries"
